+++ admin-backend/src/main/java/com/yeeiee/controller/DictTypeController.java
@@ -2,6 +2,7 @@ package com.yeeiee.controller;
 
 import com.baomidou.mybatisplus.core.conditions.query.LambdaQueryWrapper;
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
+import com.yeeiee.cache.DictCacheManager;
 import com.yeeiee.domain.entity.DictData;
 import com.yeeiee.domain.entity.DictType;
 import com.yeeiee.domain.form.DictTypeForm;
@@ -15,6 +16,8 @@ import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
 import lombok.val;
+import org.springframework.cache.annotation.CacheEvict;
+import org.springframework.util.CollectionUtils;
 import org.springframework.util.StringUtils;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -45,6 +48,7 @@ public class DictTypeController {
 
     private final DictTypeService dictTypeService;
     private final DictDataService dictDataService;
+    private final DictCacheManager dictCacheManager;
 
     @Operation(summary = "查询字典类型分页")
     @GetMapping("/{size}/{current}")
@@ -79,6 +83,7 @@ public class DictTypeController {
 
     @Operation(summary = "更新字典类型")
     @PutMapping
+    @CacheEvict(cacheNames = DictCacheManager.DICT_CACHE, key = "#dictTypeForm.id")
     public R<Void> modifyDictType(@Validated(DictTypeForm.Update.class) @RequestBody DictTypeForm dictTypeForm) {
         dictTypeService.updateById(dictTypeForm.toBean());
         return R.ok();
@@ -86,6 +91,7 @@ public class DictTypeController {
 
     @Operation(summary = "删除字典类型")
     @DeleteMapping("/{id}")
+    @CacheEvict(cacheNames = DictCacheManager.DICT_CACHE, key = "#id")
     public R<Void> removeDictTypeById(@PathVariable("id") Long id) {
         val dictDataList = dictDataService.lambdaQuery()
                 .eq(DictData::getTypeId, id)
@@ -102,15 +108,18 @@ public class DictTypeController {
     @Operation(summary = "批量删除字典类型")
     @DeleteMapping
     public R<Void> removeDictTypeByIds(@RequestParam("ids") @Parameter(description = "需要删除的id列表") Collection<Long> ids) {
-        val dictDataList = dictDataService.lambdaQuery()
-                .in(DictData::getTypeId, ids)
-                .list();
+        if (!CollectionUtils.isEmpty(ids)) {
+            val dictDataList = dictDataService.lambdaQuery()
+                    .in(DictData::getTypeId, ids)
+                    .list();
 
-        if (!dictDataList.isEmpty()) {
-            throw new DmlOperationException("删除失败，尚有字典数据依赖");
+            if (!dictDataList.isEmpty()) {
+                throw new DmlOperationException("删除失败，尚有字典数据依赖");
+            }
+
+            dictCacheManager.evictByTypeIds(ids);
+            dictTypeService.removeByIds(ids);
         }
-
-        dictTypeService.removeByIds(ids);
         return R.ok();
     }
 }
